test framework is functionally complete, all tests are stable and passing
